+++ frontend/src/main.js
@@ -8,7 +8,7 @@ import VueI18n from 'vue-i18n'
 import router from './routes'
 import "./mixin"
 import "./icons"
-import App from '@/App.vue'
+const App = () => import('@/App.vue')
 import store from './store'
 import language from './languages'